Forwards responder chain actions to the selected pane
This change checks if the selected pane is a valid responder to the proposed action. It does this by overriding `supplementalTarget(forAction:sender:)` and querying if the active pane responds to the action or itself forwards a valid responder. This necessitated a change to `PreferencesTabViewController` to expose (internally) the currently active view controller.

To test this, and a valid use case, is a font based preference. `NSFontPanel` will only send a targetless `changeFont:` action when the font changes, which means you can't hook anything up in IB, etc. to intercept this message. I've added a proof-of-concept font picker button that will launch the shared font panel and will print the font manager when the font changes to illustrate the efficacy of this change.